Added a new module 'definitions.py' which handles the 'config.json' file to avoid circular imports. Moved the function 'front_back_mapper' to 'helpers.py'.
